+++ index.html
@@ -64,8 +64,9 @@
                     <h2 data-i18n="about.title"></h2>
                     <p data-i18n="about.description"></p>
                     <p><span data-i18n="about.github"></span><a class="btn btn-primary" role="button" href="https://github.com/haddadanas/Masterclass_UHH" target="_blank" rel="noopener noreferrer"><i class="fab fa-github"></i></a></p>
-                    <p><span data-i18n="about.sponsors"></span><img src="assets/img/UHH_Universität_Hamburg_Logo.svg" alt="UHH Logo" height="70px"><img src="assets/img/desy-circle.svg" alt="DESY Logo" height="70"></p>
+                    <p><span data-i18n="about.sponsors"></span><a href="https://www.uni-hamburg.de/" target="_blank" rel="noopener noreferrer"><img src="assets/img/UHH_Universität_Hamburg_Logo.svg" title="Universität Hamburg" alt="UHH Logo" height="70px" data-bs-toggle="tooltip"></a><a href="https://www.desy.de/" target="_blank" rel="noopener noreferrer"><img src="assets/img/desy-circle.svg" title="Deutsches Elektronen-Synchrotron" alt="DESY Logo" height="70" data-bs-toggle="tooltip"></a></p>
                 </div>
+                <hr class="my-4">
                 <h2 style="text-align: center;" data-i18n="about.links"></h2>
                 <div class="row">
                     <div class="col-md-6 col-lg-4">
