+++ ts/analysis.ts
